added document api service and dtos folder
