+++ src/contexts/authContext.tsx
@@ -1,6 +1,7 @@
 import React, { createContext, useCallback, useEffect, useState } from 'react';
 import Loading from 'components/loading';
-import { authApiService, LoginDto, OperationStatus } from 'services/api/AuthApiService';
+import { authApiService } from 'services/api/AuthApiService';
+import { LoginDto, OperationStatus } from 'services/api/dtos';
 
 export interface User {
   id: string;
