+++ src/screens/monitor/hooks/useAuthentication.ts
@@ -1,0 +1,4 @@
+import { authApiService } from './../../../services/api/AuthApiService';
+import { useApiService } from "./useApiService";
+
+export const useAuthenticationApi = useApiService(authApiService);
