+++ src/services/api/AuthApiService.ts
@@ -1,56 +1,6 @@
 import { AxiosInstance, AxiosError } from 'axios';
 import { axios } from 'config/http';
-
-export interface OperationStatus<TPayload> {
-  success: boolean;
-  error?: string;
-  payload?: TPayload;
-  status?: number;
-}
-
-export interface UserDto {
-  id: string;
-  name: string;
-  surname: string;
-  role: string;
-  email: string;
-}
-
-export interface LoginDto {
-  token: string;
-  user: UserDto;
-}
-
-export interface RecoverPasswordDto {
-  userId: string;
-  token: string;
-}
-
-export enum Status {
-  NOU = 'nou',
-  IN_ANALIZA = 'in analiza',
-  REVIZUIT = 'revizuit',
-}
-
-export interface DocumentDto {
-  id: string;
-  createdAt: Date;
-  updatedAt: Date;
-  title: string;
-  project: string;
-  identificator: string;
-  publicationDate: Date;
-  source: string;
-  status: Status;
-  assignedUser?: string;
-  deadline?: Date;
-  originalFormat?: string;
-  numberOfPages?: number;
-  textInterpretationPrecision?: number;
-  numberOfIdentifiedArticles?: number;
-  numberOfIdentifiedTerms?: number;
-  attachments?: string[];
-}
+import { LoginDto, OperationStatus, RecoverPasswordDto } from './dtos';
 
 class AuthApiService {
   constructor(private readonly httpClient: AxiosInstance) {}
@@ -131,23 +81,6 @@ class AuthApiService {
       return {
         success: false,
         status: error.response?.status,
-      };
-    }
-  }
-
-  async getDocuments(): Promise<OperationStatus<Document[]>> {
-    try {
-      const response = await this.httpClient.get<Document[]>('/document');
-
-      return {
-        success: true,
-        payload: response.data,
-      }
-    } catch (err: any) {
-      const error: AxiosError = err;
-      return {
-        success: false,
-        error: error.response?.statusText,
       };
     }
   }
