+++ src/services/api/DocumentApiService.ts
@@ -1,0 +1,27 @@
+import { AxiosInstance, AxiosError } from 'axios';
+import { axios } from 'config/http';
+import { OperationStatus, DocumentDto } from './dtos';
+
+class DocumentApiService {
+  constructor(private readonly httpClient: AxiosInstance) {}
+
+  async getDocuments(): Promise<OperationStatus<DocumentDto[]>> {
+    try {
+      const response = await this.httpClient.get<DocumentDto[]>('/document');
+
+      return {
+        success: true,
+        payload: response.data,
+      }
+    } catch (err: any) {
+      const error: AxiosError = err;
+      return {
+        success: false,
+        error: error.response?.statusText,
+      };
+    }
+  }
+
+}
+
+export const documentApiService = new DocumentApiService(axios);
