+++ src/services/api/dtos/document.ts
@@ -1,0 +1,25 @@
+export enum Status {
+  NOU = 'nou',
+  IN_ANALIZA = 'in analiza',
+  REVIZUIT = 'revizuit',
+}
+
+export interface DocumentDto {
+  id: string;
+  createdAt: Date;
+  updatedAt: Date;
+  title: string;
+  project: string;
+  identificator: string;
+  publicationDate: Date;
+  source: string;
+  status: Status;
+  assignedUser?: string;
+  deadline?: Date;
+  originalFormat?: string;
+  numberOfPages?: number;
+  textInterpretationPrecision?: number;
+  numberOfIdentifiedArticles?: number;
+  numberOfIdentifiedTerms?: number;
+  attachments?: string[];
+}
