+++ src/services/api/dtos/generic.ts
@@ -1,0 +1,6 @@
+export interface OperationStatus<TPayload> {
+  success: boolean;
+  error?: string;
+  payload?: TPayload;
+  status?: number;
+}
