+++ src/services/api/dtos/index.ts
@@ -1,0 +1,3 @@
+export * from './auth';
+export * from './document';
+export * from './generic';
